add get feature pos function
